Moved the _use_fontconfig and _use_freetype config methods to gui_object
